Split array validate test to separate module
